update ProfileService relationships
